Add new unit and functional tests.
